Ref: Created /arbitrators folder and moved files inside it
