+++ app/src/components/market/sections/common_sections/arbitrators/display_arbitrator/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { Arbitrator } from '../../../../util/types'
+import { Arbitrator } from '../../../../../../util/types'
 import { ArbitratorIcon } from '../arbitrator_icon'
 
 const AWrapper = styled.a`
